Refactor symbol formatting as to not repeat code
Async-signal-safe version of symbol formatting function was added to
nativelib. Now, javalib's StackTraceElement formatting uses that
async-signal-safe function to implement its own.
