+++ javalib/src/main/scala/java/lang/StackTraceElement.scala
@@ -2,7 +2,9 @@ package java.lang
 
 import scalanative.unsafe.{CString, fromCString}
 import scalanative.libc.string.strlen
-import scalanative.runtime.Platform.isWindows
+import scalanative.unsigned._
+import scala.scalanative.unsafe._
+import scala.scalanative.runtime.SymbolFormatter
 
 final class StackTraceElement(
     val getClassName: String,
@@ -47,113 +49,15 @@ private[lang] object StackTraceElement {
   object Fail extends scala.util.control.NoStackTrace
 
   def fromSymbol(sym: CString): StackTraceElement = {
-    val len = strlen(sym)
-    var pos = 0
-    var className = ""
-    var methodName = ""
+    val className = stackalloc[CChar](1024.toUInt)
+    val methodName = stackalloc[CChar](1024.toUInt)
+    SymbolFormatter.asyncSafeFromSymbol(sym, className, methodName)
 
-    def readSymbol(): Boolean = {
-      // On Windows symbol names are different then on Unix platforms.
-      // Due to differences in implementation between WinDbg and libUnwind used
-      // on each platform, symbols on Windows do not contain '_' prefix.
-      if (!isWindows() && read() != '_') {
-        false
-      } else if (read() != 'S') {
-        false
-      } else {
-        readGlobal()
-      }
-    }
-
-    def readGlobal(): Boolean = read() match {
-      case 'M' =>
-        val id = readIdent()
-        if (id.length == 0) {
-          false
-        } else {
-          className = id
-          readSig()
-        }
-      case _ =>
-        false
-    }
-
-    def readSig(): Boolean = read() match {
-      case 'R' =>
-        methodName = "<init>"
-        true
-      case 'D' | 'P' | 'C' | 'G' =>
-        val id = readIdent()
-        if (id.length == 0) {
-          false
-        } else {
-          methodName = id
-          true
-        }
-      case 'K' =>
-        readSig()
-      case _ =>
-        false
-    }
-
-    def readIdent(): String = {
-      val n = readNumber()
-      if (n <= 0) {
-        ""
-      } else if (!inBounds(pos) || !inBounds(pos + n)) {
-        ""
-      } else {
-        val chars = new Array[Char](n)
-        var i = 0
-        while (i < n) {
-          chars(i) = sym(pos + i).toChar
-          i += 1
-        }
-        pos += n
-        new String(chars)
-      }
-    }
-
-    def readNumber(): Int = {
-      val start = pos
-      var number = 0
-      while ('0' <= at(pos) && at(pos) <= '9') {
-        number = number * 10 + (at(pos) - '0').toInt
-        pos += 1
-      }
-      if (start == pos) {
-        -1
-      } else {
-        number
-      }
-    }
-
-    def read(): Char = {
-      if (inBounds(pos)) {
-        val res = sym(pos).toChar
-        pos += 1
-        res
-      } else {
-        -1.toChar
-      }
-    }
-
-    def at(pos: Int): Char = {
-      if (inBounds(pos)) {
-        sym(pos).toChar
-      } else {
-        -1.toChar
-      }
-    }
-
-    def inBounds(pos: Int) =
-      pos >= 0 && pos < len.toLong
-
-    if (!readSymbol()) {
-      className = "<none>"
-      methodName = fromCString(sym)
-    }
-
-    new StackTraceElement(className, methodName, null, 0)
+    new StackTraceElement(
+      fromCString(className),
+      fromCString(methodName),
+      null,
+      0
+    )
   }
 }
