+++ nativelib/src/main/scala/scala/scalanative/runtime/libc.scala
@@ -12,6 +12,8 @@ object libc {
   def free(ptr: RawPtr): Unit = extern
   def strlen(str: CString): CSize = extern
   def wcslen(str: CWideString): CSize = extern
+  def strcpy(dest: CString, src: CString): CString = extern
+  def strcat(dest: CString, src: CString): CString = extern
   def memcpy(dst: RawPtr, src: RawPtr, count: CSize): RawPtr = extern
   def memcmp(lhs: RawPtr, rhs: RawPtr, count: CSize): CInt = extern
   def memset(dest: RawPtr, ch: CInt, count: CSize): RawPtr = extern
